SALE SCREEN

+++ app/src/main/java/com/example/canelinhaestoque/data/model/SaleItem.kt
@@ -3,7 +3,7 @@ package com.example.canelinhaestoque.data.model
 data class SaleItem(
     val productId: String = "",
     val name: String = "",
-    val quantity: Int = 0,
+    val quantity: Double = 0.0,
     val unitPrice: Double = 0.0,
     val itemDiscount: Double = 0.0
 )
